refactor(users): move permission helpers to users/access.py
user_has_permission, user_can_access_property, user_can_access_property_id,
and get_user_role lived in properties/access.py, making them unavailable to
the upcoming payments app without a cross-app import from properties.

Moved all four to users/access.py (the shared app). properties/access.py is
now a thin re-export shim so existing call sites in properties/views.py are
unaffected. The payments app can import directly from users.access.
